Added delete

+++ src/main/java/leadDash/leadrepository/LeadRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import leadDash.model.Lead;
 
 @Repository
-public interface LeadRepository extends CrudRepository<Lead, Long> {
+public interface LeadRepository extends CrudRepository<Lead, Integer> {
 
 }
